requirements.txt
easydict==1.9
numpy==1.19.1
PyYAML==5.3.1
scipy==1.5.2
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
